pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["dbp"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
psycopg2-binary = "^2.9.3"
black = "^22.8.0"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
Markdown = "^3.4.1"
drf-yasg = "^1.21.4"
factory-boy = "^3.2.1"
Faker = "^15.1.0"
djangorestframework-simplejwt = "^5.2.1"
coveralls = "^3.3.1"
PyYAML = "^6.0"
gunicorn = "^20.1.0"
django-on-heroku = "^1.1.2"
rollbar = "^0.16.3"
newrelic = "^8.4.0"
django-storages = "^1.13.1"
boto3 = "^1.26.5"
Pillow = "^9.3.0"
drf-extensions = "^0.7.1"

[tool.poetry.dev-dependencies]
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
freezegun = "^1.2.2"
coverage = "^6.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
